# Conflicts:
#	tkpdcore/src/main/java/com/tokopedia/core/base/di/component/AppComponent.java
#	tkpdcore/src/main/java/com/tokopedia/core/network/di/module/NetModule.java
#	tkpdcore/src/main/java/com/tokopedia/core/network/di/module/OkHttpClientModule.java
#	tkpdseller/src/main/AndroidManifest.xml
#	tkpdseller/src/main/res/values/attr.xml
#	tkpdseller/src/main/res/values/strings.xml

+++ tkpdseller/src/main/java/com/tokopedia/seller/topads/utils/ViewUtils.java
@@ -3,12 +3,17 @@ package com.tokopedia.seller.topads.utils;
 import android.annotation.TargetApi;
 import android.content.Context;
 import android.os.Build;
+import android.support.annotation.NonNull;
 import android.view.Window;
 
 import com.tokopedia.seller.R;
 import com.tokopedia.seller.myproduct.utils.CurrencyFormatter;
 import com.tokopedia.seller.topads.constant.TopAdsConstant;
 import com.tokopedia.core.network.retrofit.exception.ResponseErrorException;
+
+import java.io.IOException;
+import java.net.SocketTimeoutException;
+import java.net.UnknownHostException;
 
 /**
  * @author normansyahputa on 2/16/17.
@@ -39,7 +44,7 @@ public class ViewUtils {
         return errorMessage;
     }
 
-    public static String getClickBudgetError(Context context, float clickBudget) {
+    public static String getClickBudgetError(Context context, double clickBudget) {
         if (clickBudget < TopAdsConstant.BUDGET_MULTIPLE_BY) {
             return context.getString(R.string.error_top_ads_click_budget_minimum, String.valueOf(TopAdsConstant.BUDGET_MULTIPLE_BY));
         }
@@ -52,7 +57,7 @@ public class ViewUtils {
         return null;
     }
 
-    public static String getDailyBudgetError(Context context, float clickBudget, float dailyBudget) {
+    public static String getDailyBudgetError(Context context, float clickBudget, double dailyBudget) {
         if (dailyBudget <= 0) {
             return context.getString(R.string.error_top_ads_daily_budget_cannot_empyt);
         }
